# PERF: Series.any/all