--- a/pandas/core/dtypes/dtypes.py
+++ b/pandas/core/dtypes/dtypes.py
@@ -862,10 +862,7 @@ class PeriodDtype(PeriodDtypeBase, PandasExtensionDtype):
     _metadata = ("freq",)
     _match = re.compile(r"(P|p)eriod\[(?P<freq>.+)\]")
     _cache_dtypes: dict[str_type, PandasExtensionDtype] = {}
-    # error: Incompatible types in assignment (expression has type
-    # "Callable[[PeriodDtypeBase], int]", base class "PandasExtensionDtype"
-    # defined the type as "Callable[[PandasExtensionDtype], int]")
-    __hash__ = PeriodDtypeBase.__hash__  # type: ignore[assignment]
+    __hash__ = PeriodDtypeBase.__hash__
 
     def __new__(cls, freq):
         """