- [x] closes #26032 (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

Example based on OP from #26032
```
import numpy as np
import pandas as pd
ser = pd.Series(np.random.randint(0, 2, 100000)).astype(bool)

In [2]: %timeit s.any(skipna=True)
5.79 µs ± 719 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- PR
8.44 µs ± 473 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- main
```

We'll be hard-pressed to get trim any more overhead from this, so I'm declaring this as closing #26032.